pipe-hr/requirements.txt
supabase
python-dotenv
tenacity